python/requirements.txt
openai==1.66.3
packaging==23.2
python-dotenv==1.0.0
selenium==4.16.0
robotframework
robotframework-seleniumlibrary
robotframework-pythonlibcore
grpcio==1.62.0
grpcio-tools==1.62.0
protobuf==4.25.3
appdirs==1.4.4 